Split SvgTransformCollection
